Replaced Registry path expanded by dfvfs Windows path resolver.
