+++ plaso/dfwinreg/registry.py
@@ -1,12 +1,6 @@
 # -*- coding: utf-8 -*-
 """Classes for Windows Registry access."""
 
-import logging
-
-from dfvfs.helpers import windows_path_resolver
-from dfvfs.resolver import resolver
-
-from plaso.dfwinreg import interface
 from plaso.dfwinreg import regf
 
 
@@ -430,164 +424,5 @@ class WinRegistry(object):
     except IOError:
       file_object.close()
       registry_file = None
-
-    return registry_file
-
-
-class PathSpecWinRegistryFileReader(interface.WinRegistryFileReader):
-  """A single path specification Windows Registry file reader."""
-
-  def __init__(self, path_spec):
-    """Initializes a Windows Registry file reader object.
-
-    Args:
-      path_spec: a path specification (instance of dfvfs.PathSpec).
-    """
-    super(PathSpecWinRegistryFileReader, self).__init__()
-    self._path_spec = path_spec
-
-  def _OpenPathSpec(self, path_spec, ascii_codepage=u'cp1252'):
-    """Opens the Windows Registry file specified by the path specification.
-
-    Args:
-      path_spec: a path specification (instance of dfvfs.PathSpec).
-      ascii_codepage: optional ASCII string codepage. The default is cp1252
-                      (or windows-1252).
-
-    Returns:
-      The Windows Registry file (instance of WinRegistryFile) or None.
-    """
-    if not path_spec:
-      return
-
-    file_entry = resolver.Resolver.OpenFileEntry(path_spec)
-    if file_entry is None:
-      return
-
-    file_object = file_entry.GetFileObject()
-    if file_object is None:
-      return
-
-    registry_file = regf.REGFWinRegistryFile(ascii_codepage=ascii_codepage)
-    try:
-      registry_file.Open(file_object)
-    except IOError as exception:
-      logging.warning(
-          u'Unable to open Windows Registry file with error: {0:s}'.format(
-              exception))
-      file_object.close()
-      return
-
-    return registry_file
-
-  def Open(self, path, ascii_codepage=u'cp1252'):
-    """Opens the Windows Registry file specified by the path.
-
-    Args:
-      path: the path of the Windows Registry file.
-      ascii_codepage: optional ASCII string codepage. The default is cp1252
-                      (or windows-1252).
-
-    Returns:
-      The Windows Registry file (instance of WinRegistryFile) or None.
-    """
-    return self._OpenPathSpec(self._path_spec)
-
-
-class SearcherWinRegistryFileReader(interface.WinRegistryFileReader):
-  """A file system searcher-based Windows Registry file reader."""
-
-  def __init__(self, file_system, mount_point, path_attributes=None):
-    """Initializes a Windows Registry file reader object.
-
-    Args:
-      file_system: the file system object (instance of vfs.FileSystem).
-      mount_point: the mount point path specification (instance of
-                   path.PathSpec).
-      path_attributes: optional dictionary of path attributes.
-    """
-    super(SearcherWinRegistryFileReader, self).__init__()
-    self._path_resolver = windows_path_resolver.WindowsPathResolver(
-        file_system, mount_point)
-
-    if path_attributes:
-      for attribute_name, attribute_value in iter(path_attributes.items()):
-        # TODO: fix the call to this class and make sure only relevant
-        # values are passed.
-        if attribute_name == u'systemroot':
-          self._path_resolver.SetEnvironmentVariable(
-              u'SystemRoot', attribute_value)
-
-  def _OpenPathSpec(self, path_spec, ascii_codepage=u'cp1252'):
-    """Opens the Windows Registry file specified by the path specification.
-
-    Args:
-      path_spec: a path specification (instance of dfvfs.PathSpec).
-      ascii_codepage: optional ASCII string codepage. The default is cp1252
-                      (or windows-1252).
-
-    Returns:
-      The Windows Registry file (instance of WinRegistryFile) or None.
-    """
-    if not path_spec:
-      return
-
-    file_object = resolver.Resolver.OpenFileObject(path_spec)
-    if file_object is None:
-      return
-
-    # TODO: fix empty file object.
-    registry_file = regf.REGFWinRegistryFile(ascii_codepage=ascii_codepage)
-    try:
-      registry_file.Open(file_object)
-    except IOError as exception:
-      logging.warning(
-          u'Unable to open Windows Registry file with error: {0:s}'.format(
-              exception))
-      file_object.close()
-      return
-
-    return registry_file
-
-  def Open(self, path, ascii_codepage=u'cp1252'):
-    """Opens the Windows Registry file specified by the path.
-
-    Args:
-      path: the path of the Windows Registry file.
-      ascii_codepage: optional ASCII string codepage. The default is cp1252
-                      (or windows-1252).
-
-    Returns:
-      The Windows Registry file (instance of WinRegistryFile) or None.
-    """
-    path_spec = self._path_resolver.ResolvePath(path)
-    if path_spec is None:
-      return
-
-    return self._OpenPathSpec(path_spec)
-
-
-class FileObjectWinRegistryFileReader(interface.WinRegistryFileReader):
-  """A single file-like object Windows Registry file reader."""
-
-  def Open(self, file_object, ascii_codepage=u'cp1252'):
-    """Opens a Windows Registry file-like object.
-
-    Args:
-      file_object: the Windows Registry file-like object.
-      ascii_codepage: optional ASCII string codepage. The default is cp1252
-                      (or windows-1252).
-
-    Returns:
-      The Windows Registry file (instance of WinRegistryFile) or None.
-    """
-    registry_file = regf.REGFWinRegistryFile(ascii_codepage=ascii_codepage)
-    try:
-      registry_file.Open(file_object)
-    except IOError as exception:
-      logging.warning(
-          u'Unable to open Windows Registry file with error: {0:s}'.format(
-              exception))
-      return
 
     return registry_file
